requirements.txt
appnope==0.1.0
backcall==0.1.0
cycler==0.10.0
decorator==4.4.0
imageio==2.5.0
ipython==7.6.0
ipython-genutils==0.2.0
jedi==0.14.0
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.0
networkx==2.3
numpy==1.16.4
opencv-contrib-python==4.1.0.25
pandas==0.24.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pip==19.1.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
setuptools==41.0.1
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
wheel==0.33.4
